requirements.txt
streamlit
Pillow
onnxruntime
numpy
